fet(bindings): updated bindings configuration
