+++ scripts/Cargo.toml
@@ -20,8 +20,9 @@ log = "0.4"
 env_logger = "0.11.5"
 ethers = "2.0.7"
 
-[[bin]]
-name = "generate_bindings"
-path = "src/build.rs"
 
-
+[build-dependencies]
+ethers = "2.0.7"
+cainome = { git = "https://github.com/cartridge-gg/cainome", tag = "v0.2.9", features = [
+    "abigen-rs",
+] }
