+++ scripts/src/bind/address_whitelist.rs
@@ -1,10 +1,10 @@
 #[derive(Debug)]
-pub struct finder<A: starknet::accounts::ConnectedAccount + Sync> {
+pub struct address_whitelist<A: starknet::accounts::ConnectedAccount + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub account: A,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
     pub fn new(address: starknet::core::types::FieldElement, account: A) -> Self {
         Self {
             address,
@@ -25,12 +25,12 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
     }
 }
 #[derive(Debug)]
-pub struct finderReader<P: starknet::providers::Provider + Sync> {
+pub struct address_whitelistReader<P: starknet::providers::Provider + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub provider: P,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<P: starknet::providers::Provider + Sync> finderReader<P> {
+impl<P: starknet::providers::Provider + Sync> address_whitelistReader<P> {
     pub fn new(address: starknet::core::types::FieldElement, provider: P) -> Self {
         Self {
             address,
@@ -48,72 +48,6 @@ impl<P: starknet::providers::Provider + Sync> finderReader<P> {
     }
     pub fn set_block(mut self, block_id: starknet::core::types::BlockId) {
         self.block_id = block_id;
-    }
-}
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct InterfaceImplementationChanged {
-    pub interface_name: starknet::core::types::FieldElement,
-    pub new_implementation_address: cainome::cairo_serde::ContractAddress,
-}
-impl cainome::cairo_serde::CairoSerde for InterfaceImplementationChanged {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        let mut __size = 0;
-        __size
-            += starknet::core::types::FieldElement::cairo_serialized_size(
-                &__rust.interface_name,
-            );
-        __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.new_implementation_address,
-            );
-        __size
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out
-            .extend(
-                starknet::core::types::FieldElement::cairo_serialize(
-                    &__rust.interface_name,
-                ),
-            );
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    &__rust.new_implementation_address,
-                ),
-            );
-        __out
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let mut __offset = __offset;
-        let interface_name = starknet::core::types::FieldElement::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += starknet::core::types::FieldElement::cairo_serialized_size(
-                &interface_name,
-            );
-        let new_implementation_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &new_implementation_address,
-            );
-        Ok(InterfaceImplementationChanged {
-            interface_name,
-            new_implementation_address,
-        })
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
@@ -179,6 +113,96 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub struct RemovedFromWhitelist {
+    pub removed_address: cainome::cairo_serde::ContractAddress,
+}
+impl cainome::cairo_serde::CairoSerde for RemovedFromWhitelist {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        let mut __size = 0;
+        __size
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.removed_address,
+            );
+        __size
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    &__rust.removed_address,
+                ),
+            );
+        __out
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let mut __offset = __offset;
+        let removed_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &removed_address,
+            );
+        Ok(RemovedFromWhitelist {
+            removed_address,
+        })
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub struct AddedToWhitelist {
+    pub added_address: cainome::cairo_serde::ContractAddress,
+}
+impl cainome::cairo_serde::CairoSerde for AddedToWhitelist {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        let mut __size = 0;
+        __size
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.added_address,
+            );
+        __size
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    &__rust.added_address,
+                ),
+            );
+        __out
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let mut __offset = __offset;
+        let added_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &added_address,
+            );
+        Ok(AddedToWhitelist { added_address })
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
 pub struct OwnershipTransferStarted {
     pub previous_owner: cainome::cairo_serde::ContractAddress,
     pub new_owner: cainome::cairo_serde::ContractAddress,
@@ -241,20 +265,13 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferStarted {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {
-    InterfaceImplementationChanged(InterfaceImplementationChanged),
-    OwnableEvent(Event),
-}
+pub enum Event {}
 impl cainome::cairo_serde::CairoSerde for Event {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::InterfaceImplementationChanged(val) => {
-                InterfaceImplementationChanged::cairo_serialized_size(val) + 1
-            }
-            Event::OwnableEvent(val) => Event::cairo_serialized_size(val) + 1,
             _ => 0,
         }
     }
@@ -262,15 +279,86 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::InterfaceImplementationChanged(val) => {
+            _ => vec![],
+        }
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let __index: u128 = __felts[__offset].try_into().unwrap();
+        match __index as usize {
+            _ => {
+                return Err(
+                    cainome::cairo_serde::Error::Deserialize(
+                        format!("Index not handle for enum {}", "Event"),
+                    ),
+                );
+            }
+        }
+    }
+}
+impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+    type Error = String;
+    fn try_from(
+        event: starknet::core::types::EmittedEvent,
+    ) -> Result<Self, Self::Error> {
+        use cainome::cairo_serde::CairoSerde;
+        if event.keys.is_empty() {
+            return Err("Event has no key".to_string());
+        }
+        Err(format!("Could not match any event from keys {:?}", event.keys))
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub enum Event {
+    AddedToWhitelist(AddedToWhitelist),
+    RemovedFromWhitelist(RemovedFromWhitelist),
+    OwnableEvent(OwnableCptEvent),
+    ReentrancyGuardEvent(Event),
+}
+impl cainome::cairo_serde::CairoSerde for Event {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        match __rust {
+            Event::AddedToWhitelist(val) => {
+                AddedToWhitelist::cairo_serialized_size(val) + 1
+            }
+            Event::RemovedFromWhitelist(val) => {
+                RemovedFromWhitelist::cairo_serialized_size(val) + 1
+            }
+            Event::OwnableEvent(val) => OwnableCptEvent::cairo_serialized_size(val) + 1,
+            Event::ReentrancyGuardEvent(val) => Event::cairo_serialized_size(val) + 1,
+            _ => 0,
+        }
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        match __rust {
+            Event::AddedToWhitelist(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
-                temp.extend(InterfaceImplementationChanged::cairo_serialize(val));
+                temp.extend(AddedToWhitelist::cairo_serialize(val));
+                temp
+            }
+            Event::RemovedFromWhitelist(val) => {
+                let mut temp = vec![];
+                temp.extend(usize::cairo_serialize(&1usize));
+                temp.extend(RemovedFromWhitelist::cairo_serialize(val));
                 temp
             }
             Event::OwnableEvent(val) => {
                 let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&1usize));
+                temp.extend(usize::cairo_serialize(&2usize));
+                temp.extend(OwnableCptEvent::cairo_serialize(val));
+                temp
+            }
+            Event::ReentrancyGuardEvent(val) => {
+                let mut temp = vec![];
+                temp.extend(usize::cairo_serialize(&3usize));
                 temp.extend(Event::cairo_serialize(val));
                 temp
             }
@@ -285,16 +373,31 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::InterfaceImplementationChanged(
-                        InterfaceImplementationChanged::cairo_deserialize(
-                            __felts,
-                            __offset + 1,
-                        )?,
+                    Event::AddedToWhitelist(
+                        AddedToWhitelist::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
             }
             1usize => {
-                Ok(Event::OwnableEvent(Event::cairo_deserialize(__felts, __offset + 1)?))
+                Ok(
+                    Event::RemovedFromWhitelist(
+                        RemovedFromWhitelist::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
+            }
+            2usize => {
+                Ok(
+                    Event::OwnableEvent(
+                        OwnableCptEvent::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
+            }
+            3usize => {
+                Ok(
+                    Event::ReentrancyGuardEvent(
+                        Event::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
             }
             _ => {
                 return Err(
@@ -317,16 +420,14 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         }
         let selector = event.keys[0];
         if selector
-            == starknet::core::utils::get_selector_from_name(
-                    "InterfaceImplementationChanged",
-                )
+            == starknet::core::utils::get_selector_from_name("AddedToWhitelist")
                 .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "InterfaceImplementationChanged")
+                    panic!("Invalid selector for {}", "AddedToWhitelist")
                 })
         {
             let mut key_offset = 0 + 1;
             let mut data_offset = 0;
-            let interface_name = match starknet::core::types::FieldElement::cairo_deserialize(
+            let added_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
                 &event.data,
                 data_offset,
             ) {
@@ -335,28 +436,8 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     return Err(
                         format!(
                             "Could not deserialize field {} for {}: {:?}",
-                            "interface_name",
-                            "InterfaceImplementationChanged",
-                            e,
-                        ),
-                    );
-                }
-            };
-            data_offset
-                += starknet::core::types::FieldElement::cairo_serialized_size(
-                    &interface_name,
-                );
-            let new_implementation_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.data,
-                data_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "new_implementation_address",
-                            "InterfaceImplementationChanged",
+                            "added_address",
+                            "AddedToWhitelist",
                             e,
                         ),
                     );
@@ -364,12 +445,42 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
             };
             data_offset
                 += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &new_implementation_address,
+                    &added_address,
+                );
+            return Ok(Event::AddedToWhitelist(AddedToWhitelist { added_address }));
+        }
+        let selector = event.keys[0];
+        if selector
+            == starknet::core::utils::get_selector_from_name("RemovedFromWhitelist")
+                .unwrap_or_else(|_| {
+                    panic!("Invalid selector for {}", "RemovedFromWhitelist")
+                })
+        {
+            let mut key_offset = 0 + 1;
+            let mut data_offset = 0;
+            let removed_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.data,
+                data_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "removed_address",
+                            "RemovedFromWhitelist",
+                            e,
+                        ),
+                    );
+                }
+            };
+            data_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &removed_address,
                 );
             return Ok(
-                Event::InterfaceImplementationChanged(InterfaceImplementationChanged {
-                    interface_name,
-                    new_implementation_address,
+                Event::RemovedFromWhitelist(RemovedFromWhitelist {
+                    removed_address,
                 }),
             );
         }
@@ -424,7 +535,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferred(OwnershipTransferred {
+                    OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                         previous_owner,
                         new_owner,
                     }),
@@ -482,7 +593,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                    OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                         previous_owner,
                         new_owner,
                     }),
@@ -493,20 +604,20 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {
+pub enum OwnableCptEvent {
     OwnershipTransferred(OwnershipTransferred),
     OwnershipTransferStarted(OwnershipTransferStarted),
 }
-impl cainome::cairo_serde::CairoSerde for Event {
+impl cainome::cairo_serde::CairoSerde for OwnableCptEvent {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 OwnershipTransferred::cairo_serialized_size(val) + 1
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 OwnershipTransferStarted::cairo_serialized_size(val) + 1
             }
             _ => 0,
@@ -516,13 +627,13 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
                 temp.extend(OwnershipTransferred::cairo_serialize(val));
                 temp
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&1usize));
                 temp.extend(OwnershipTransferStarted::cairo_serialize(val));
@@ -539,14 +650,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::OwnershipTransferred(
+                    OwnableCptEvent::OwnershipTransferred(
                         OwnershipTransferred::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
             }
             1usize => {
                 Ok(
-                    Event::OwnershipTransferStarted(
+                    OwnableCptEvent::OwnershipTransferStarted(
                         OwnershipTransferStarted::cairo_deserialize(
                             __felts,
                             __offset + 1,
@@ -557,14 +668,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
             _ => {
                 return Err(
                     cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
+                        format!("Index not handle for enum {}", "OwnableCptEvent"),
                     ),
                 );
             }
         }
     }
 }
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+impl TryFrom<starknet::core::types::EmittedEvent> for OwnableCptEvent {
     type Error = String;
     fn try_from(
         event: starknet::core::types::EmittedEvent,
@@ -623,7 +734,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferred(OwnershipTransferred {
+                OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                     previous_owner,
                     new_owner,
                 }),
@@ -679,7 +790,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                     previous_owner,
                     new_owner,
                 }),
@@ -688,7 +799,43 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         Err(format!("Could not match any event from keys {:?}", event.keys))
     }
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
+    #[allow(clippy::ptr_arg)]
+    #[allow(clippy::too_many_arguments)]
+    pub fn is_on_whitelist(
+        &self,
+        element_to_check: &cainome::cairo_serde::ContractAddress,
+    ) -> cainome::cairo_serde::call::FCall<A::Provider, bool> {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_check),
+            );
+        let __call = starknet::core::types::FunctionCall {
+            contract_address: self.address,
+            entry_point_selector: starknet::macros::selector!("is_on_whitelist"),
+            calldata: __calldata,
+        };
+        cainome::cairo_serde::call::FCall::new(__call, self.provider())
+    }
+    #[allow(clippy::ptr_arg)]
+    #[allow(clippy::too_many_arguments)]
+    pub fn get_whitelist(
+        &self,
+    ) -> cainome::cairo_serde::call::FCall<
+        A::Provider,
+        Vec<cainome::cairo_serde::ContractAddress>,
+    > {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        let __call = starknet::core::types::FunctionCall {
+            contract_address: self.address,
+            entry_point_selector: starknet::macros::selector!("get_whitelist"),
+            calldata: __calldata,
+        };
+        cainome::cairo_serde::call::FCall::new(__call, self.provider())
+    }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
     pub fn owner(
@@ -708,27 +855,71 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn get_implementation_address(
+    pub fn add_to_whitelist_getcall(
         &self,
-        interface_name: &starknet::core::types::FieldElement,
-    ) -> cainome::cairo_serde::call::FCall<
-        A::Provider,
-        cainome::cairo_serde::ContractAddress,
-    > {
+        new_element: &cainome::cairo_serde::ContractAddress,
+    ) -> starknet::accounts::Call {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(new_element));
+        starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("add_to_whitelist"),
+            calldata: __calldata,
+        }
+    }
+    #[allow(clippy::ptr_arg)]
+    pub fn add_to_whitelist(
+        &self,
+        new_element: &cainome::cairo_serde::ContractAddress,
+    ) -> starknet::accounts::Execution<A> {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(new_element));
+        let __call = starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("add_to_whitelist"),
+            calldata: __calldata,
+        };
+        self.account.execute(vec![__call])
+    }
+    #[allow(clippy::ptr_arg)]
+    #[allow(clippy::too_many_arguments)]
+    pub fn remove_from_whitelist_getcall(
+        &self,
+        element_to_remove: &cainome::cairo_serde::ContractAddress,
+    ) -> starknet::accounts::Call {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
             .extend(
-                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_remove),
             );
-        let __call = starknet::core::types::FunctionCall {
-            contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!(
-                "get_implementation_address"
-            ),
+        starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("remove_from_whitelist"),
+            calldata: __calldata,
+        }
+    }
+    #[allow(clippy::ptr_arg)]
+    pub fn remove_from_whitelist(
+        &self,
+        element_to_remove: &cainome::cairo_serde::ContractAddress,
+    ) -> starknet::accounts::Execution<A> {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_remove),
+            );
+        let __call = starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("remove_from_whitelist"),
             calldata: __calldata,
         };
-        cainome::cairo_serde::call::FCall::new(__call, self.provider())
+        self.account.execute(vec![__call])
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
@@ -784,58 +975,44 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
         };
         self.account.execute(vec![__call])
     }
+}
+impl<P: starknet::providers::Provider + Sync> address_whitelistReader<P> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn change_implementation_address_getcall(
+    pub fn is_on_whitelist(
         &self,
-        interface_name: &starknet::core::types::FieldElement,
-        implementation_address: &cainome::cairo_serde::ContractAddress,
-    ) -> starknet::accounts::Call {
+        element_to_check: &cainome::cairo_serde::ContractAddress,
+    ) -> cainome::cairo_serde::call::FCall<P, bool> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
             .extend(
-                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_check),
             );
-        __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    implementation_address,
-                ),
-            );
-        starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("change_implementation_address"),
-            calldata: __calldata,
-        }
-    }
-    #[allow(clippy::ptr_arg)]
-    pub fn change_implementation_address(
-        &self,
-        interface_name: &starknet::core::types::FieldElement,
-        implementation_address: &cainome::cairo_serde::ContractAddress,
-    ) -> starknet::accounts::Execution<A> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(
-                starknet::core::types::FieldElement::cairo_serialize(interface_name),
-            );
-        __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    implementation_address,
-                ),
-            );
-        let __call = starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("change_implementation_address"),
+        let __call = starknet::core::types::FunctionCall {
+            contract_address: self.address,
+            entry_point_selector: starknet::macros::selector!("is_on_whitelist"),
             calldata: __calldata,
         };
-        self.account.execute(vec![__call])
+        cainome::cairo_serde::call::FCall::new(__call, self.provider())
     }
-}
-impl<P: starknet::providers::Provider + Sync> finderReader<P> {
+    #[allow(clippy::ptr_arg)]
+    #[allow(clippy::too_many_arguments)]
+    pub fn get_whitelist(
+        &self,
+    ) -> cainome::cairo_serde::call::FCall<
+        P,
+        Vec<cainome::cairo_serde::ContractAddress>,
+    > {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        let __call = starknet::core::types::FunctionCall {
+            contract_address: self.address,
+            entry_point_selector: starknet::macros::selector!("get_whitelist"),
+            calldata: __calldata,
+        };
+        cainome::cairo_serde::call::FCall::new(__call, self.provider())
+    }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
     pub fn owner(
@@ -846,27 +1023,6 @@ impl<P: starknet::providers::Provider + Sync> finderReader<P> {
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
             entry_point_selector: starknet::macros::selector!("owner"),
-            calldata: __calldata,
-        };
-        cainome::cairo_serde::call::FCall::new(__call, self.provider())
-    }
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::too_many_arguments)]
-    pub fn get_implementation_address(
-        &self,
-        interface_name: &starknet::core::types::FieldElement,
-    ) -> cainome::cairo_serde::call::FCall<P, cainome::cairo_serde::ContractAddress> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(
-                starknet::core::types::FieldElement::cairo_serialize(interface_name),
-            );
-        let __call = starknet::core::types::FunctionCall {
-            contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!(
-                "get_implementation_address"
-            ),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
