+++ scripts/src/bind/finder.rs
@@ -1,10 +1,10 @@
 #[derive(Debug)]
-pub struct identifier_whitelist<A: starknet::accounts::ConnectedAccount + Sync> {
+pub struct finder<A: starknet::accounts::ConnectedAccount + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub account: A,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
     pub fn new(address: starknet::core::types::FieldElement, account: A) -> Self {
         Self {
             address,
@@ -25,12 +25,12 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
     }
 }
 #[derive(Debug)]
-pub struct identifier_whitelistReader<P: starknet::providers::Provider + Sync> {
+pub struct finderReader<P: starknet::providers::Provider + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub provider: P,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<P: starknet::providers::Provider + Sync> identifier_whitelistReader<P> {
+impl<P: starknet::providers::Provider + Sync> finderReader<P> {
     pub fn new(address: starknet::core::types::FieldElement, provider: P) -> Self {
         Self {
             address,
@@ -51,18 +51,23 @@ impl<P: starknet::providers::Provider + Sync> identifier_whitelistReader<P> {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct SupportedIdentifierRemoved {
-    pub identifier: starknet::core::types::FieldElement,
+pub struct OwnershipTransferStarted {
+    pub previous_owner: cainome::cairo_serde::ContractAddress,
+    pub new_owner: cainome::cairo_serde::ContractAddress,
 }
-impl cainome::cairo_serde::CairoSerde for SupportedIdentifierRemoved {
+impl cainome::cairo_serde::CairoSerde for OwnershipTransferStarted {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         let mut __size = 0;
         __size
-            += starknet::core::types::FieldElement::cairo_serialized_size(
-                &__rust.identifier,
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.previous_owner,
+            );
+        __size
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.new_owner,
             );
         __size
     }
@@ -72,7 +77,13 @@ impl cainome::cairo_serde::CairoSerde for SupportedIdentifierRemoved {
         let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
         __out
             .extend(
-                starknet::core::types::FieldElement::cairo_serialize(&__rust.identifier),
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    &__rust.previous_owner,
+                ),
+            );
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(&__rust.new_owner),
             );
         __out
     }
@@ -81,14 +92,89 @@ impl cainome::cairo_serde::CairoSerde for SupportedIdentifierRemoved {
         __offset: usize,
     ) -> cainome::cairo_serde::Result<Self::RustType> {
         let mut __offset = __offset;
-        let identifier = starknet::core::types::FieldElement::cairo_deserialize(
+        let previous_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
             __felts,
             __offset,
         )?;
         __offset
-            += starknet::core::types::FieldElement::cairo_serialized_size(&identifier);
-        Ok(SupportedIdentifierRemoved {
-            identifier,
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &previous_owner,
+            );
+        let new_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(&new_owner);
+        Ok(OwnershipTransferStarted {
+            previous_owner,
+            new_owner,
+        })
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub struct InterfaceImplementationChanged {
+    pub interface_name: starknet::core::types::FieldElement,
+    pub new_implementation_address: cainome::cairo_serde::ContractAddress,
+}
+impl cainome::cairo_serde::CairoSerde for InterfaceImplementationChanged {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        let mut __size = 0;
+        __size
+            += starknet::core::types::FieldElement::cairo_serialized_size(
+                &__rust.interface_name,
+            );
+        __size
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.new_implementation_address,
+            );
+        __size
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
+        __out
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(
+                    &__rust.interface_name,
+                ),
+            );
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    &__rust.new_implementation_address,
+                ),
+            );
+        __out
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let mut __offset = __offset;
+        let interface_name = starknet::core::types::FieldElement::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += starknet::core::types::FieldElement::cairo_serialized_size(
+                &interface_name,
+            );
+        let new_implementation_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &new_implementation_address,
+            );
+        Ok(InterfaceImplementationChanged {
+            interface_name,
+            new_implementation_address,
         })
     }
 }
@@ -155,114 +241,9 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct SupportedIdentifierAdded {
-    pub identifier: starknet::core::types::FieldElement,
-}
-impl cainome::cairo_serde::CairoSerde for SupportedIdentifierAdded {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        let mut __size = 0;
-        __size
-            += starknet::core::types::FieldElement::cairo_serialized_size(
-                &__rust.identifier,
-            );
-        __size
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out
-            .extend(
-                starknet::core::types::FieldElement::cairo_serialize(&__rust.identifier),
-            );
-        __out
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let mut __offset = __offset;
-        let identifier = starknet::core::types::FieldElement::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += starknet::core::types::FieldElement::cairo_serialized_size(&identifier);
-        Ok(SupportedIdentifierAdded {
-            identifier,
-        })
-    }
-}
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct OwnershipTransferStarted {
-    pub previous_owner: cainome::cairo_serde::ContractAddress,
-    pub new_owner: cainome::cairo_serde::ContractAddress,
-}
-impl cainome::cairo_serde::CairoSerde for OwnershipTransferStarted {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        let mut __size = 0;
-        __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.previous_owner,
-            );
-        __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.new_owner,
-            );
-        __size
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    &__rust.previous_owner,
-                ),
-            );
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(&__rust.new_owner),
-            );
-        __out
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let mut __offset = __offset;
-        let previous_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &previous_owner,
-            );
-        let new_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(&new_owner);
-        Ok(OwnershipTransferStarted {
-            previous_owner,
-            new_owner,
-        })
-    }
-}
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
 pub enum Event {
-    SupportedIdentifierAdded(SupportedIdentifierAdded),
-    SupportedIdentifierRemoved(SupportedIdentifierRemoved),
-    OwnableEvent(Event),
+    InterfaceImplementationChanged(InterfaceImplementationChanged),
+    OwnableEvent(OwnableCptEvent),
 }
 impl cainome::cairo_serde::CairoSerde for Event {
     type RustType = Self;
@@ -270,13 +251,10 @@ impl cainome::cairo_serde::CairoSerde for Event {
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::SupportedIdentifierAdded(val) => {
-                SupportedIdentifierAdded::cairo_serialized_size(val) + 1
+            Event::InterfaceImplementationChanged(val) => {
+                InterfaceImplementationChanged::cairo_serialized_size(val) + 1
             }
-            Event::SupportedIdentifierRemoved(val) => {
-                SupportedIdentifierRemoved::cairo_serialized_size(val) + 1
-            }
-            Event::OwnableEvent(val) => Event::cairo_serialized_size(val) + 1,
+            Event::OwnableEvent(val) => OwnableCptEvent::cairo_serialized_size(val) + 1,
             _ => 0,
         }
     }
@@ -284,22 +262,16 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::SupportedIdentifierAdded(val) => {
+            Event::InterfaceImplementationChanged(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
-                temp.extend(SupportedIdentifierAdded::cairo_serialize(val));
-                temp
-            }
-            Event::SupportedIdentifierRemoved(val) => {
-                let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&1usize));
-                temp.extend(SupportedIdentifierRemoved::cairo_serialize(val));
+                temp.extend(InterfaceImplementationChanged::cairo_serialize(val));
                 temp
             }
             Event::OwnableEvent(val) => {
                 let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&2usize));
-                temp.extend(Event::cairo_serialize(val));
+                temp.extend(usize::cairo_serialize(&1usize));
+                temp.extend(OwnableCptEvent::cairo_serialize(val));
                 temp
             }
             _ => vec![],
@@ -313,8 +285,8 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::SupportedIdentifierAdded(
-                        SupportedIdentifierAdded::cairo_deserialize(
+                    Event::InterfaceImplementationChanged(
+                        InterfaceImplementationChanged::cairo_deserialize(
                             __felts,
                             __offset + 1,
                         )?,
@@ -323,16 +295,10 @@ impl cainome::cairo_serde::CairoSerde for Event {
             }
             1usize => {
                 Ok(
-                    Event::SupportedIdentifierRemoved(
-                        SupportedIdentifierRemoved::cairo_deserialize(
-                            __felts,
-                            __offset + 1,
-                        )?,
+                    Event::OwnableEvent(
+                        OwnableCptEvent::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
-            }
-            2usize => {
-                Ok(Event::OwnableEvent(Event::cairo_deserialize(__felts, __offset + 1)?))
             }
             _ => {
                 return Err(
@@ -355,51 +321,16 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         }
         let selector = event.keys[0];
         if selector
-            == starknet::core::utils::get_selector_from_name("SupportedIdentifierAdded")
-                .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "SupportedIdentifierAdded")
-                })
-        {
-            let mut key_offset = 0 + 1;
-            let mut data_offset = 0;
-            let identifier = match starknet::core::types::FieldElement::cairo_deserialize(
-                &event.data,
-                data_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "identifier",
-                            "SupportedIdentifierAdded",
-                            e,
-                        ),
-                    );
-                }
-            };
-            data_offset
-                += starknet::core::types::FieldElement::cairo_serialized_size(
-                    &identifier,
-                );
-            return Ok(
-                Event::SupportedIdentifierAdded(SupportedIdentifierAdded {
-                    identifier,
-                }),
-            );
-        }
-        let selector = event.keys[0];
-        if selector
             == starknet::core::utils::get_selector_from_name(
-                    "SupportedIdentifierRemoved",
+                    "InterfaceImplementationChanged",
                 )
                 .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "SupportedIdentifierRemoved")
+                    panic!("Invalid selector for {}", "InterfaceImplementationChanged")
                 })
         {
             let mut key_offset = 0 + 1;
             let mut data_offset = 0;
-            let identifier = match starknet::core::types::FieldElement::cairo_deserialize(
+            let interface_name = match starknet::core::types::FieldElement::cairo_deserialize(
                 &event.data,
                 data_offset,
             ) {
@@ -408,8 +339,8 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     return Err(
                         format!(
                             "Could not deserialize field {} for {}: {:?}",
-                            "identifier",
-                            "SupportedIdentifierRemoved",
+                            "interface_name",
+                            "InterfaceImplementationChanged",
                             e,
                         ),
                     );
@@ -417,11 +348,32 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
             };
             data_offset
                 += starknet::core::types::FieldElement::cairo_serialized_size(
-                    &identifier,
+                    &interface_name,
+                );
+            let new_implementation_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.data,
+                data_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "new_implementation_address",
+                            "InterfaceImplementationChanged",
+                            e,
+                        ),
+                    );
+                }
+            };
+            data_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &new_implementation_address,
                 );
             return Ok(
-                Event::SupportedIdentifierRemoved(SupportedIdentifierRemoved {
-                    identifier,
+                Event::InterfaceImplementationChanged(InterfaceImplementationChanged {
+                    interface_name,
+                    new_implementation_address,
                 }),
             );
         }
@@ -476,7 +428,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferred(OwnershipTransferred {
+                    OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                         previous_owner,
                         new_owner,
                     }),
@@ -534,7 +486,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                    OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                         previous_owner,
                         new_owner,
                     }),
@@ -545,20 +497,20 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {
+pub enum OwnableCptEvent {
     OwnershipTransferred(OwnershipTransferred),
     OwnershipTransferStarted(OwnershipTransferStarted),
 }
-impl cainome::cairo_serde::CairoSerde for Event {
+impl cainome::cairo_serde::CairoSerde for OwnableCptEvent {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 OwnershipTransferred::cairo_serialized_size(val) + 1
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 OwnershipTransferStarted::cairo_serialized_size(val) + 1
             }
             _ => 0,
@@ -568,13 +520,13 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
                 temp.extend(OwnershipTransferred::cairo_serialize(val));
                 temp
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&1usize));
                 temp.extend(OwnershipTransferStarted::cairo_serialize(val));
@@ -591,14 +543,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::OwnershipTransferred(
+                    OwnableCptEvent::OwnershipTransferred(
                         OwnershipTransferred::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
             }
             1usize => {
                 Ok(
-                    Event::OwnershipTransferStarted(
+                    OwnableCptEvent::OwnershipTransferStarted(
                         OwnershipTransferStarted::cairo_deserialize(
                             __felts,
                             __offset + 1,
@@ -609,14 +561,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
             _ => {
                 return Err(
                     cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
+                        format!("Index not handle for enum {}", "OwnableCptEvent"),
                     ),
                 );
             }
         }
     }
 }
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+impl TryFrom<starknet::core::types::EmittedEvent> for OwnableCptEvent {
     type Error = String;
     fn try_from(
         event: starknet::core::types::EmittedEvent,
@@ -675,7 +627,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferred(OwnershipTransferred {
+                OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                     previous_owner,
                     new_owner,
                 }),
@@ -731,7 +683,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                     previous_owner,
                     new_owner,
                 }),
@@ -740,20 +692,27 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         Err(format!("Could not match any event from keys {:?}", event.keys))
     }
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> finder<A> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn is_identifier_supported(
+    pub fn get_implementation_address(
         &self,
-        identifier: &starknet::core::types::FieldElement,
-    ) -> cainome::cairo_serde::call::FCall<A::Provider, bool> {
+        interface_name: &starknet::core::types::FieldElement,
+    ) -> cainome::cairo_serde::call::FCall<
+        A::Provider,
+        cainome::cairo_serde::ContractAddress,
+    > {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+            );
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("is_identifier_supported"),
+            entry_point_selector: starknet::macros::selector!(
+                "get_implementation_address"
+            ),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
@@ -777,64 +736,50 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn add_supported_identifier_getcall(
+    pub fn change_implementation_address_getcall(
         &self,
-        identifier: &starknet::core::types::FieldElement,
+        interface_name: &starknet::core::types::FieldElement,
+        implementation_address: &cainome::cairo_serde::ContractAddress,
     ) -> starknet::accounts::Call {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+            );
+        __calldata
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    implementation_address,
+                ),
+            );
         starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("add_supported_identifier"),
+            selector: starknet::macros::selector!("change_implementation_address"),
             calldata: __calldata,
         }
     }
     #[allow(clippy::ptr_arg)]
-    pub fn add_supported_identifier(
+    pub fn change_implementation_address(
         &self,
-        identifier: &starknet::core::types::FieldElement,
+        interface_name: &starknet::core::types::FieldElement,
+        implementation_address: &cainome::cairo_serde::ContractAddress,
     ) -> starknet::accounts::Execution<A> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+            );
+        __calldata
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    implementation_address,
+                ),
+            );
         let __call = starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("add_supported_identifier"),
-            calldata: __calldata,
-        };
-        self.account.execute(vec![__call])
-    }
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::too_many_arguments)]
-    pub fn remove_supported_identifier_getcall(
-        &self,
-        identifier: &starknet::core::types::FieldElement,
-    ) -> starknet::accounts::Call {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
-        starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("remove_supported_identifier"),
-            calldata: __calldata,
-        }
-    }
-    #[allow(clippy::ptr_arg)]
-    pub fn remove_supported_identifier(
-        &self,
-        identifier: &starknet::core::types::FieldElement,
-    ) -> starknet::accounts::Execution<A> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
-        let __call = starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("remove_supported_identifier"),
+            selector: starknet::macros::selector!("change_implementation_address"),
             calldata: __calldata,
         };
         self.account.execute(vec![__call])
@@ -894,20 +839,24 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
         self.account.execute(vec![__call])
     }
 }
-impl<P: starknet::providers::Provider + Sync> identifier_whitelistReader<P> {
+impl<P: starknet::providers::Provider + Sync> finderReader<P> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn is_identifier_supported(
+    pub fn get_implementation_address(
         &self,
-        identifier: &starknet::core::types::FieldElement,
-    ) -> cainome::cairo_serde::call::FCall<P, bool> {
+        interface_name: &starknet::core::types::FieldElement,
+    ) -> cainome::cairo_serde::call::FCall<P, cainome::cairo_serde::ContractAddress> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(interface_name),
+            );
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("is_identifier_supported"),
+            entry_point_selector: starknet::macros::selector!(
+                "get_implementation_address"
+            ),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
