+++ scripts/src/bind/identifier_whitelist.rs
@@ -1,10 +1,10 @@
 #[derive(Debug)]
-pub struct store<A: starknet::accounts::ConnectedAccount + Sync> {
+pub struct identifier_whitelist<A: starknet::accounts::ConnectedAccount + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub account: A,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
     pub fn new(address: starknet::core::types::FieldElement, account: A) -> Self {
         Self {
             address,
@@ -25,12 +25,12 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
     }
 }
 #[derive(Debug)]
-pub struct storeReader<P: starknet::providers::Provider + Sync> {
+pub struct identifier_whitelistReader<P: starknet::providers::Provider + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub provider: P,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<P: starknet::providers::Provider + Sync> storeReader<P> {
+impl<P: starknet::providers::Provider + Sync> identifier_whitelistReader<P> {
     pub fn new(address: starknet::core::types::FieldElement, provider: P) -> Self {
         Self {
             address,
@@ -51,23 +51,18 @@ impl<P: starknet::providers::Provider + Sync> storeReader<P> {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct OwnershipTransferred {
-    pub previous_owner: cainome::cairo_serde::ContractAddress,
-    pub new_owner: cainome::cairo_serde::ContractAddress,
+pub struct SupportedIdentifierAdded {
+    pub identifier: starknet::core::types::FieldElement,
 }
-impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
+impl cainome::cairo_serde::CairoSerde for SupportedIdentifierAdded {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         let mut __size = 0;
         __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.previous_owner,
-            );
-        __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.new_owner,
+            += starknet::core::types::FieldElement::cairo_serialized_size(
+                &__rust.identifier,
             );
         __size
     }
@@ -77,13 +72,7 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
         let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
         __out
             .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    &__rust.previous_owner,
-                ),
-            );
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(&__rust.new_owner),
+                starknet::core::types::FieldElement::cairo_serialize(&__rust.identifier),
             );
         __out
     }
@@ -92,23 +81,56 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
         __offset: usize,
     ) -> cainome::cairo_serde::Result<Self::RustType> {
         let mut __offset = __offset;
-        let previous_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+        let identifier = starknet::core::types::FieldElement::cairo_deserialize(
             __felts,
             __offset,
         )?;
         __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &previous_owner,
+            += starknet::core::types::FieldElement::cairo_serialized_size(&identifier);
+        Ok(SupportedIdentifierAdded {
+            identifier,
+        })
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub struct SupportedIdentifierRemoved {
+    pub identifier: starknet::core::types::FieldElement,
+}
+impl cainome::cairo_serde::CairoSerde for SupportedIdentifierRemoved {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        let mut __size = 0;
+        __size
+            += starknet::core::types::FieldElement::cairo_serialized_size(
+                &__rust.identifier,
             );
-        let new_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+        __size
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
+        __out
+            .extend(
+                starknet::core::types::FieldElement::cairo_serialize(&__rust.identifier),
+            );
+        __out
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let mut __offset = __offset;
+        let identifier = starknet::core::types::FieldElement::cairo_deserialize(
             __felts,
             __offset,
         )?;
         __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(&new_owner);
-        Ok(OwnershipTransferred {
-            previous_owner,
-            new_owner,
+            += starknet::core::types::FieldElement::cairo_serialized_size(&identifier);
+        Ok(SupportedIdentifierRemoved {
+            identifier,
         })
     }
 }
@@ -175,24 +197,40 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferStarted {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct NewFinalFee {
-    pub new_final_fee: cainome::cairo_serde::U256,
+pub struct OwnershipTransferred {
+    pub previous_owner: cainome::cairo_serde::ContractAddress,
+    pub new_owner: cainome::cairo_serde::ContractAddress,
 }
-impl cainome::cairo_serde::CairoSerde for NewFinalFee {
+impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         let mut __size = 0;
         __size
-            += cainome::cairo_serde::U256::cairo_serialized_size(&__rust.new_final_fee);
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.previous_owner,
+            );
+        __size
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &__rust.new_owner,
+            );
         __size
     }
     fn cairo_serialize(
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out.extend(cainome::cairo_serde::U256::cairo_serialize(&__rust.new_final_fee));
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(
+                    &__rust.previous_owner,
+                ),
+            );
+        __out
+            .extend(
+                cainome::cairo_serde::ContractAddress::cairo_serialize(&__rust.new_owner),
+            );
         __out
     }
     fn cairo_deserialize(
@@ -200,29 +238,41 @@ impl cainome::cairo_serde::CairoSerde for NewFinalFee {
         __offset: usize,
     ) -> cainome::cairo_serde::Result<Self::RustType> {
         let mut __offset = __offset;
-        let new_final_fee = cainome::cairo_serde::U256::cairo_deserialize(
+        let previous_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
             __felts,
             __offset,
         )?;
-        __offset += cainome::cairo_serde::U256::cairo_serialized_size(&new_final_fee);
-        Ok(NewFinalFee { new_final_fee })
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                &previous_owner,
+            );
+        let new_owner = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+            __felts,
+            __offset,
+        )?;
+        __offset
+            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(&new_owner);
+        Ok(OwnershipTransferred {
+            previous_owner,
+            new_owner,
+        })
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {
+pub enum OwnableCptEvent {
     OwnershipTransferred(OwnershipTransferred),
     OwnershipTransferStarted(OwnershipTransferStarted),
 }
-impl cainome::cairo_serde::CairoSerde for Event {
+impl cainome::cairo_serde::CairoSerde for OwnableCptEvent {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 OwnershipTransferred::cairo_serialized_size(val) + 1
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 OwnershipTransferStarted::cairo_serialized_size(val) + 1
             }
             _ => 0,
@@ -232,13 +282,13 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
                 temp.extend(OwnershipTransferred::cairo_serialize(val));
                 temp
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&1usize));
                 temp.extend(OwnershipTransferStarted::cairo_serialize(val));
@@ -255,14 +305,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::OwnershipTransferred(
+                    OwnableCptEvent::OwnershipTransferred(
                         OwnershipTransferred::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
             }
             1usize => {
                 Ok(
-                    Event::OwnershipTransferStarted(
+                    OwnableCptEvent::OwnershipTransferStarted(
                         OwnershipTransferStarted::cairo_deserialize(
                             __felts,
                             __offset + 1,
@@ -273,14 +323,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
             _ => {
                 return Err(
                     cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
+                        format!("Index not handle for enum {}", "OwnableCptEvent"),
                     ),
                 );
             }
         }
     }
 }
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+impl TryFrom<starknet::core::types::EmittedEvent> for OwnableCptEvent {
     type Error = String;
     fn try_from(
         event: starknet::core::types::EmittedEvent,
@@ -339,7 +389,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferred(OwnershipTransferred {
+                OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                     previous_owner,
                     new_owner,
                 }),
@@ -395,7 +445,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                     previous_owner,
                     new_owner,
                 }),
@@ -406,8 +456,9 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
 pub enum Event {
-    NewFinalFee(NewFinalFee),
-    OwnableEvent(Event),
+    SupportedIdentifierAdded(SupportedIdentifierAdded),
+    SupportedIdentifierRemoved(SupportedIdentifierRemoved),
+    OwnableEvent(OwnableCptEvent),
 }
 impl cainome::cairo_serde::CairoSerde for Event {
     type RustType = Self;
@@ -415,8 +466,13 @@ impl cainome::cairo_serde::CairoSerde for Event {
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::NewFinalFee(val) => NewFinalFee::cairo_serialized_size(val) + 1,
-            Event::OwnableEvent(val) => Event::cairo_serialized_size(val) + 1,
+            Event::SupportedIdentifierAdded(val) => {
+                SupportedIdentifierAdded::cairo_serialized_size(val) + 1
+            }
+            Event::SupportedIdentifierRemoved(val) => {
+                SupportedIdentifierRemoved::cairo_serialized_size(val) + 1
+            }
+            Event::OwnableEvent(val) => OwnableCptEvent::cairo_serialized_size(val) + 1,
             _ => 0,
         }
     }
@@ -424,16 +480,22 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::NewFinalFee(val) => {
+            Event::SupportedIdentifierAdded(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
-                temp.extend(NewFinalFee::cairo_serialize(val));
+                temp.extend(SupportedIdentifierAdded::cairo_serialize(val));
+                temp
+            }
+            Event::SupportedIdentifierRemoved(val) => {
+                let mut temp = vec![];
+                temp.extend(usize::cairo_serialize(&1usize));
+                temp.extend(SupportedIdentifierRemoved::cairo_serialize(val));
                 temp
             }
             Event::OwnableEvent(val) => {
                 let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&1usize));
-                temp.extend(Event::cairo_serialize(val));
+                temp.extend(usize::cairo_serialize(&2usize));
+                temp.extend(OwnableCptEvent::cairo_serialize(val));
                 temp
             }
             _ => vec![],
@@ -447,13 +509,30 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::NewFinalFee(
-                        NewFinalFee::cairo_deserialize(__felts, __offset + 1)?,
+                    Event::SupportedIdentifierAdded(
+                        SupportedIdentifierAdded::cairo_deserialize(
+                            __felts,
+                            __offset + 1,
+                        )?,
                     ),
                 )
             }
             1usize => {
-                Ok(Event::OwnableEvent(Event::cairo_deserialize(__felts, __offset + 1)?))
+                Ok(
+                    Event::SupportedIdentifierRemoved(
+                        SupportedIdentifierRemoved::cairo_deserialize(
+                            __felts,
+                            __offset + 1,
+                        )?,
+                    ),
+                )
+            }
+            2usize => {
+                Ok(
+                    Event::OwnableEvent(
+                        OwnableCptEvent::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
             }
             _ => {
                 return Err(
@@ -476,12 +555,14 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         }
         let selector = event.keys[0];
         if selector
-            == starknet::core::utils::get_selector_from_name("NewFinalFee")
-                .unwrap_or_else(|_| panic!("Invalid selector for {}", "NewFinalFee"))
+            == starknet::core::utils::get_selector_from_name("SupportedIdentifierAdded")
+                .unwrap_or_else(|_| {
+                    panic!("Invalid selector for {}", "SupportedIdentifierAdded")
+                })
         {
             let mut key_offset = 0 + 1;
             let mut data_offset = 0;
-            let new_final_fee = match cainome::cairo_serde::U256::cairo_deserialize(
+            let identifier = match starknet::core::types::FieldElement::cairo_deserialize(
                 &event.data,
                 data_offset,
             ) {
@@ -490,16 +571,59 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     return Err(
                         format!(
                             "Could not deserialize field {} for {}: {:?}",
-                            "new_final_fee",
-                            "NewFinalFee",
+                            "identifier",
+                            "SupportedIdentifierAdded",
                             e,
                         ),
                     );
                 }
             };
             data_offset
-                += cainome::cairo_serde::U256::cairo_serialized_size(&new_final_fee);
-            return Ok(Event::NewFinalFee(NewFinalFee { new_final_fee }));
+                += starknet::core::types::FieldElement::cairo_serialized_size(
+                    &identifier,
+                );
+            return Ok(
+                Event::SupportedIdentifierAdded(SupportedIdentifierAdded {
+                    identifier,
+                }),
+            );
+        }
+        let selector = event.keys[0];
+        if selector
+            == starknet::core::utils::get_selector_from_name(
+                    "SupportedIdentifierRemoved",
+                )
+                .unwrap_or_else(|_| {
+                    panic!("Invalid selector for {}", "SupportedIdentifierRemoved")
+                })
+        {
+            let mut key_offset = 0 + 1;
+            let mut data_offset = 0;
+            let identifier = match starknet::core::types::FieldElement::cairo_deserialize(
+                &event.data,
+                data_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "identifier",
+                            "SupportedIdentifierRemoved",
+                            e,
+                        ),
+                    );
+                }
+            };
+            data_offset
+                += starknet::core::types::FieldElement::cairo_serialized_size(
+                    &identifier,
+                );
+            return Ok(
+                Event::SupportedIdentifierRemoved(SupportedIdentifierRemoved {
+                    identifier,
+                }),
+            );
         }
         let selector = event.keys[0];
         if selector
@@ -552,7 +676,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferred(OwnershipTransferred {
+                    OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                         previous_owner,
                         new_owner,
                     }),
@@ -610,7 +734,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                 );
             return Ok(
                 Event::OwnableEvent(
-                    Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                    OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                         previous_owner,
                         new_owner,
                     }),
@@ -620,41 +744,20 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         Err(format!("Could not match any event from keys {:?}", event.keys))
     }
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> identifier_whitelist<A> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn pay_oracle_fees(
+    pub fn is_identifier_supported(
         &self,
-        erc20_address: &cainome::cairo_serde::ContractAddress,
-        amount: &cainome::cairo_serde::U256,
-    ) -> cainome::cairo_serde::call::FCall<A::Provider, ()> {
+        identifier: &starknet::core::types::FieldElement,
+    ) -> cainome::cairo_serde::call::FCall<A::Provider, bool> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(erc20_address),
-            );
-        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(amount));
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("pay_oracle_fees"),
-            calldata: __calldata,
-        };
-        cainome::cairo_serde::call::FCall::new(__call, self.provider())
-    }
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::too_many_arguments)]
-    pub fn compute_final_fee(
-        &self,
-        currency: &cainome::cairo_serde::ContractAddress,
-    ) -> cainome::cairo_serde::call::FCall<A::Provider, cainome::cairo_serde::U256> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
-        let __call = starknet::core::types::FunctionCall {
-            contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("compute_final_fee"),
+            entry_point_selector: starknet::macros::selector!("is_identifier_supported"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
@@ -678,36 +781,64 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn set_final_fee_getcall(
+    pub fn add_supported_identifier_getcall(
         &self,
-        currency: &cainome::cairo_serde::ContractAddress,
-        new_final_fee: &cainome::cairo_serde::U256,
+        identifier: &starknet::core::types::FieldElement,
     ) -> starknet::accounts::Call {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
-        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(new_final_fee));
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
         starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("set_final_fee"),
+            selector: starknet::macros::selector!("add_supported_identifier"),
             calldata: __calldata,
         }
     }
     #[allow(clippy::ptr_arg)]
-    pub fn set_final_fee(
+    pub fn add_supported_identifier(
         &self,
-        currency: &cainome::cairo_serde::ContractAddress,
-        new_final_fee: &cainome::cairo_serde::U256,
+        identifier: &starknet::core::types::FieldElement,
     ) -> starknet::accounts::Execution<A> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
-        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(new_final_fee));
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
         let __call = starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("set_final_fee"),
+            selector: starknet::macros::selector!("add_supported_identifier"),
+            calldata: __calldata,
+        };
+        self.account.execute(vec![__call])
+    }
+    #[allow(clippy::ptr_arg)]
+    #[allow(clippy::too_many_arguments)]
+    pub fn remove_supported_identifier_getcall(
+        &self,
+        identifier: &starknet::core::types::FieldElement,
+    ) -> starknet::accounts::Call {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+        starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("remove_supported_identifier"),
+            calldata: __calldata,
+        }
+    }
+    #[allow(clippy::ptr_arg)]
+    pub fn remove_supported_identifier(
+        &self,
+        identifier: &starknet::core::types::FieldElement,
+    ) -> starknet::accounts::Execution<A> {
+        use cainome::cairo_serde::CairoSerde;
+        let mut __calldata = vec![];
+        __calldata
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
+        let __call = starknet::accounts::Call {
+            to: self.address,
+            selector: starknet::macros::selector!("remove_supported_identifier"),
             calldata: __calldata,
         };
         self.account.execute(vec![__call])
@@ -767,41 +898,20 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
         self.account.execute(vec![__call])
     }
 }
-impl<P: starknet::providers::Provider + Sync> storeReader<P> {
+impl<P: starknet::providers::Provider + Sync> identifier_whitelistReader<P> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn pay_oracle_fees(
+    pub fn is_identifier_supported(
         &self,
-        erc20_address: &cainome::cairo_serde::ContractAddress,
-        amount: &cainome::cairo_serde::U256,
-    ) -> cainome::cairo_serde::call::FCall<P, ()> {
+        identifier: &starknet::core::types::FieldElement,
+    ) -> cainome::cairo_serde::call::FCall<P, bool> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(erc20_address),
-            );
-        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(amount));
+            .extend(starknet::core::types::FieldElement::cairo_serialize(identifier));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("pay_oracle_fees"),
-            calldata: __calldata,
-        };
-        cainome::cairo_serde::call::FCall::new(__call, self.provider())
-    }
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::too_many_arguments)]
-    pub fn compute_final_fee(
-        &self,
-        currency: &cainome::cairo_serde::ContractAddress,
-    ) -> cainome::cairo_serde::call::FCall<P, cainome::cairo_serde::U256> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
-        let __call = starknet::core::types::FunctionCall {
-            contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("compute_final_fee"),
+            entry_point_selector: starknet::macros::selector!("is_identifier_supported"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
