+++ scripts/src/bind/store.rs
@@ -1,10 +1,10 @@
 #[derive(Debug)]
-pub struct address_whitelist<A: starknet::accounts::ConnectedAccount + Sync> {
+pub struct store<A: starknet::accounts::ConnectedAccount + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub account: A,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
     pub fn new(address: starknet::core::types::FieldElement, account: A) -> Self {
         Self {
             address,
@@ -25,12 +25,12 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
     }
 }
 #[derive(Debug)]
-pub struct address_whitelistReader<P: starknet::providers::Provider + Sync> {
+pub struct storeReader<P: starknet::providers::Provider + Sync> {
     pub address: starknet::core::types::FieldElement,
     pub provider: P,
     pub block_id: starknet::core::types::BlockId,
 }
-impl<P: starknet::providers::Provider + Sync> address_whitelistReader<P> {
+impl<P: starknet::providers::Provider + Sync> storeReader<P> {
     pub fn new(address: starknet::core::types::FieldElement, provider: P) -> Self {
         Self {
             address,
@@ -51,31 +51,24 @@ impl<P: starknet::providers::Provider + Sync> address_whitelistReader<P> {
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct AddedToWhitelist {
-    pub added_address: cainome::cairo_serde::ContractAddress,
+pub struct NewFinalFee {
+    pub new_final_fee: cainome::cairo_serde::U256,
 }
-impl cainome::cairo_serde::CairoSerde for AddedToWhitelist {
+impl cainome::cairo_serde::CairoSerde for NewFinalFee {
     type RustType = Self;
     const SERIALIZED_SIZE: std::option::Option<usize> = None;
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         let mut __size = 0;
         __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.added_address,
-            );
+            += cainome::cairo_serde::U256::cairo_serialized_size(&__rust.new_final_fee);
         __size
     }
     fn cairo_serialize(
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    &__rust.added_address,
-                ),
-            );
+        __out.extend(cainome::cairo_serde::U256::cairo_serialize(&__rust.new_final_fee));
         __out
     }
     fn cairo_deserialize(
@@ -83,15 +76,12 @@ impl cainome::cairo_serde::CairoSerde for AddedToWhitelist {
         __offset: usize,
     ) -> cainome::cairo_serde::Result<Self::RustType> {
         let mut __offset = __offset;
-        let added_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
+        let new_final_fee = cainome::cairo_serde::U256::cairo_deserialize(
             __felts,
             __offset,
         )?;
-        __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &added_address,
-            );
-        Ok(AddedToWhitelist { added_address })
+        __offset += cainome::cairo_serde::U256::cairo_serialized_size(&new_final_fee);
+        Ok(NewFinalFee { new_final_fee })
     }
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
@@ -153,52 +143,6 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferred {
         Ok(OwnershipTransferred {
             previous_owner,
             new_owner,
-        })
-    }
-}
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub struct RemovedFromWhitelist {
-    pub removed_address: cainome::cairo_serde::ContractAddress,
-}
-impl cainome::cairo_serde::CairoSerde for RemovedFromWhitelist {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        let mut __size = 0;
-        __size
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &__rust.removed_address,
-            );
-        __size
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        let mut __out: Vec<starknet::core::types::FieldElement> = vec![];
-        __out
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(
-                    &__rust.removed_address,
-                ),
-            );
-        __out
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let mut __offset = __offset;
-        let removed_address = cainome::cairo_serde::ContractAddress::cairo_deserialize(
-            __felts,
-            __offset,
-        )?;
-        __offset
-            += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                &removed_address,
-            );
-        Ok(RemovedFromWhitelist {
-            removed_address,
         })
     }
 }
@@ -266,8 +210,8 @@ impl cainome::cairo_serde::CairoSerde for OwnershipTransferStarted {
 }
 #[derive(Debug, PartialEq, PartialOrd, Clone)]
 pub enum Event {
-    OwnershipTransferred(OwnershipTransferred),
-    OwnershipTransferStarted(OwnershipTransferStarted),
+    NewFinalFee(NewFinalFee),
+    OwnableEvent(OwnableCptEvent),
 }
 impl cainome::cairo_serde::CairoSerde for Event {
     type RustType = Self;
@@ -275,10 +219,230 @@ impl cainome::cairo_serde::CairoSerde for Event {
     #[inline]
     fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            Event::NewFinalFee(val) => NewFinalFee::cairo_serialized_size(val) + 1,
+            Event::OwnableEvent(val) => OwnableCptEvent::cairo_serialized_size(val) + 1,
+            _ => 0,
+        }
+    }
+    fn cairo_serialize(
+        __rust: &Self::RustType,
+    ) -> Vec<starknet::core::types::FieldElement> {
+        match __rust {
+            Event::NewFinalFee(val) => {
+                let mut temp = vec![];
+                temp.extend(usize::cairo_serialize(&0usize));
+                temp.extend(NewFinalFee::cairo_serialize(val));
+                temp
+            }
+            Event::OwnableEvent(val) => {
+                let mut temp = vec![];
+                temp.extend(usize::cairo_serialize(&1usize));
+                temp.extend(OwnableCptEvent::cairo_serialize(val));
+                temp
+            }
+            _ => vec![],
+        }
+    }
+    fn cairo_deserialize(
+        __felts: &[starknet::core::types::FieldElement],
+        __offset: usize,
+    ) -> cainome::cairo_serde::Result<Self::RustType> {
+        let __index: u128 = __felts[__offset].try_into().unwrap();
+        match __index as usize {
+            0usize => {
+                Ok(
+                    Event::NewFinalFee(
+                        NewFinalFee::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
+            }
+            1usize => {
+                Ok(
+                    Event::OwnableEvent(
+                        OwnableCptEvent::cairo_deserialize(__felts, __offset + 1)?,
+                    ),
+                )
+            }
+            _ => {
+                return Err(
+                    cainome::cairo_serde::Error::Deserialize(
+                        format!("Index not handle for enum {}", "Event"),
+                    ),
+                );
+            }
+        }
+    }
+}
+impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+    type Error = String;
+    fn try_from(
+        event: starknet::core::types::EmittedEvent,
+    ) -> Result<Self, Self::Error> {
+        use cainome::cairo_serde::CairoSerde;
+        if event.keys.is_empty() {
+            return Err("Event has no key".to_string());
+        }
+        let selector = event.keys[0];
+        if selector
+            == starknet::core::utils::get_selector_from_name("NewFinalFee")
+                .unwrap_or_else(|_| panic!("Invalid selector for {}", "NewFinalFee"))
+        {
+            let mut key_offset = 0 + 1;
+            let mut data_offset = 0;
+            let new_final_fee = match cainome::cairo_serde::U256::cairo_deserialize(
+                &event.data,
+                data_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "new_final_fee",
+                            "NewFinalFee",
+                            e,
+                        ),
+                    );
+                }
+            };
+            data_offset
+                += cainome::cairo_serde::U256::cairo_serialized_size(&new_final_fee);
+            return Ok(Event::NewFinalFee(NewFinalFee { new_final_fee }));
+        }
+        let selector = event.keys[0];
+        if selector
+            == starknet::core::utils::get_selector_from_name("OwnershipTransferred")
+                .unwrap_or_else(|_| {
+                    panic!("Invalid selector for {}", "OwnershipTransferred")
+                })
+        {
+            let mut key_offset = 0 + 1;
+            let mut data_offset = 0;
+            let previous_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.keys,
+                key_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "previous_owner",
+                            "OwnershipTransferred",
+                            e,
+                        ),
+                    );
+                }
+            };
+            key_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &previous_owner,
+                );
+            let new_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.keys,
+                key_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "new_owner",
+                            "OwnershipTransferred",
+                            e,
+                        ),
+                    );
+                }
+            };
+            key_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &new_owner,
+                );
+            return Ok(
+                Event::OwnableEvent(
+                    OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
+                        previous_owner,
+                        new_owner,
+                    }),
+                ),
+            );
+        }
+        let selector = event.keys[0];
+        if selector
+            == starknet::core::utils::get_selector_from_name("OwnershipTransferStarted")
+                .unwrap_or_else(|_| {
+                    panic!("Invalid selector for {}", "OwnershipTransferStarted")
+                })
+        {
+            let mut key_offset = 0 + 1;
+            let mut data_offset = 0;
+            let previous_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.keys,
+                key_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "previous_owner",
+                            "OwnershipTransferStarted",
+                            e,
+                        ),
+                    );
+                }
+            };
+            key_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &previous_owner,
+                );
+            let new_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
+                &event.keys,
+                key_offset,
+            ) {
+                Ok(v) => v,
+                Err(e) => {
+                    return Err(
+                        format!(
+                            "Could not deserialize field {} for {}: {:?}",
+                            "new_owner",
+                            "OwnershipTransferStarted",
+                            e,
+                        ),
+                    );
+                }
+            };
+            key_offset
+                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
+                    &new_owner,
+                );
+            return Ok(
+                Event::OwnableEvent(
+                    OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
+                        previous_owner,
+                        new_owner,
+                    }),
+                ),
+            );
+        }
+        Err(format!("Could not match any event from keys {:?}", event.keys))
+    }
+}
+#[derive(Debug, PartialEq, PartialOrd, Clone)]
+pub enum OwnableCptEvent {
+    OwnershipTransferred(OwnershipTransferred),
+    OwnershipTransferStarted(OwnershipTransferStarted),
+}
+impl cainome::cairo_serde::CairoSerde for OwnableCptEvent {
+    type RustType = Self;
+    const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
+    #[inline]
+    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
+        match __rust {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 OwnershipTransferred::cairo_serialized_size(val) + 1
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 OwnershipTransferStarted::cairo_serialized_size(val) + 1
             }
             _ => 0,
@@ -288,13 +452,13 @@ impl cainome::cairo_serde::CairoSerde for Event {
         __rust: &Self::RustType,
     ) -> Vec<starknet::core::types::FieldElement> {
         match __rust {
-            Event::OwnershipTransferred(val) => {
+            OwnableCptEvent::OwnershipTransferred(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&0usize));
                 temp.extend(OwnershipTransferred::cairo_serialize(val));
                 temp
             }
-            Event::OwnershipTransferStarted(val) => {
+            OwnableCptEvent::OwnershipTransferStarted(val) => {
                 let mut temp = vec![];
                 temp.extend(usize::cairo_serialize(&1usize));
                 temp.extend(OwnershipTransferStarted::cairo_serialize(val));
@@ -311,14 +475,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
         match __index as usize {
             0usize => {
                 Ok(
-                    Event::OwnershipTransferred(
+                    OwnableCptEvent::OwnershipTransferred(
                         OwnershipTransferred::cairo_deserialize(__felts, __offset + 1)?,
                     ),
                 )
             }
             1usize => {
                 Ok(
-                    Event::OwnershipTransferStarted(
+                    OwnableCptEvent::OwnershipTransferStarted(
                         OwnershipTransferStarted::cairo_deserialize(
                             __felts,
                             __offset + 1,
@@ -329,14 +493,14 @@ impl cainome::cairo_serde::CairoSerde for Event {
             _ => {
                 return Err(
                     cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
+                        format!("Index not handle for enum {}", "OwnableCptEvent"),
                     ),
                 );
             }
         }
     }
 }
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
+impl TryFrom<starknet::core::types::EmittedEvent> for OwnableCptEvent {
     type Error = String;
     fn try_from(
         event: starknet::core::types::EmittedEvent,
@@ -395,7 +559,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferred(OwnershipTransferred {
+                OwnableCptEvent::OwnershipTransferred(OwnershipTransferred {
                     previous_owner,
                     new_owner,
                 }),
@@ -451,7 +615,7 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
                     &new_owner,
                 );
             return Ok(
-                Event::OwnershipTransferStarted(OwnershipTransferStarted {
+                OwnableCptEvent::OwnershipTransferStarted(OwnershipTransferStarted {
                     previous_owner,
                     new_owner,
                 }),
@@ -460,374 +624,41 @@ impl TryFrom<starknet::core::types::EmittedEvent> for Event {
         Err(format!("Could not match any event from keys {:?}", event.keys))
     }
 }
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {}
-impl cainome::cairo_serde::CairoSerde for Event {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        match __rust {
-            _ => 0,
-        }
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        match __rust {
-            _ => vec![],
-        }
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let __index: u128 = __felts[__offset].try_into().unwrap();
-        match __index as usize {
-            _ => {
-                return Err(
-                    cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
-                    ),
-                );
-            }
-        }
-    }
-}
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
-    type Error = String;
-    fn try_from(
-        event: starknet::core::types::EmittedEvent,
-    ) -> Result<Self, Self::Error> {
-        use cainome::cairo_serde::CairoSerde;
-        if event.keys.is_empty() {
-            return Err("Event has no key".to_string());
-        }
-        Err(format!("Could not match any event from keys {:?}", event.keys))
-    }
-}
-#[derive(Debug, PartialEq, PartialOrd, Clone)]
-pub enum Event {
-    AddedToWhitelist(AddedToWhitelist),
-    RemovedFromWhitelist(RemovedFromWhitelist),
-    OwnableEvent(Event),
-    ReentrancyGuardEvent(Event),
-}
-impl cainome::cairo_serde::CairoSerde for Event {
-    type RustType = Self;
-    const SERIALIZED_SIZE: std::option::Option<usize> = std::option::Option::None;
-    #[inline]
-    fn cairo_serialized_size(__rust: &Self::RustType) -> usize {
-        match __rust {
-            Event::AddedToWhitelist(val) => {
-                AddedToWhitelist::cairo_serialized_size(val) + 1
-            }
-            Event::RemovedFromWhitelist(val) => {
-                RemovedFromWhitelist::cairo_serialized_size(val) + 1
-            }
-            Event::OwnableEvent(val) => Event::cairo_serialized_size(val) + 1,
-            Event::ReentrancyGuardEvent(val) => Event::cairo_serialized_size(val) + 1,
-            _ => 0,
-        }
-    }
-    fn cairo_serialize(
-        __rust: &Self::RustType,
-    ) -> Vec<starknet::core::types::FieldElement> {
-        match __rust {
-            Event::AddedToWhitelist(val) => {
-                let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&0usize));
-                temp.extend(AddedToWhitelist::cairo_serialize(val));
-                temp
-            }
-            Event::RemovedFromWhitelist(val) => {
-                let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&1usize));
-                temp.extend(RemovedFromWhitelist::cairo_serialize(val));
-                temp
-            }
-            Event::OwnableEvent(val) => {
-                let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&2usize));
-                temp.extend(Event::cairo_serialize(val));
-                temp
-            }
-            Event::ReentrancyGuardEvent(val) => {
-                let mut temp = vec![];
-                temp.extend(usize::cairo_serialize(&3usize));
-                temp.extend(Event::cairo_serialize(val));
-                temp
-            }
-            _ => vec![],
-        }
-    }
-    fn cairo_deserialize(
-        __felts: &[starknet::core::types::FieldElement],
-        __offset: usize,
-    ) -> cainome::cairo_serde::Result<Self::RustType> {
-        let __index: u128 = __felts[__offset].try_into().unwrap();
-        match __index as usize {
-            0usize => {
-                Ok(
-                    Event::AddedToWhitelist(
-                        AddedToWhitelist::cairo_deserialize(__felts, __offset + 1)?,
-                    ),
-                )
-            }
-            1usize => {
-                Ok(
-                    Event::RemovedFromWhitelist(
-                        RemovedFromWhitelist::cairo_deserialize(__felts, __offset + 1)?,
-                    ),
-                )
-            }
-            2usize => {
-                Ok(Event::OwnableEvent(Event::cairo_deserialize(__felts, __offset + 1)?))
-            }
-            3usize => {
-                Ok(
-                    Event::ReentrancyGuardEvent(
-                        Event::cairo_deserialize(__felts, __offset + 1)?,
-                    ),
-                )
-            }
-            _ => {
-                return Err(
-                    cainome::cairo_serde::Error::Deserialize(
-                        format!("Index not handle for enum {}", "Event"),
-                    ),
-                );
-            }
-        }
-    }
-}
-impl TryFrom<starknet::core::types::EmittedEvent> for Event {
-    type Error = String;
-    fn try_from(
-        event: starknet::core::types::EmittedEvent,
-    ) -> Result<Self, Self::Error> {
-        use cainome::cairo_serde::CairoSerde;
-        if event.keys.is_empty() {
-            return Err("Event has no key".to_string());
-        }
-        let selector = event.keys[0];
-        if selector
-            == starknet::core::utils::get_selector_from_name("AddedToWhitelist")
-                .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "AddedToWhitelist")
-                })
-        {
-            let mut key_offset = 0 + 1;
-            let mut data_offset = 0;
-            let added_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.data,
-                data_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "added_address",
-                            "AddedToWhitelist",
-                            e,
-                        ),
-                    );
-                }
-            };
-            data_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &added_address,
-                );
-            return Ok(Event::AddedToWhitelist(AddedToWhitelist { added_address }));
-        }
-        let selector = event.keys[0];
-        if selector
-            == starknet::core::utils::get_selector_from_name("RemovedFromWhitelist")
-                .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "RemovedFromWhitelist")
-                })
-        {
-            let mut key_offset = 0 + 1;
-            let mut data_offset = 0;
-            let removed_address = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.data,
-                data_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "removed_address",
-                            "RemovedFromWhitelist",
-                            e,
-                        ),
-                    );
-                }
-            };
-            data_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &removed_address,
-                );
-            return Ok(
-                Event::RemovedFromWhitelist(RemovedFromWhitelist {
-                    removed_address,
-                }),
-            );
-        }
-        let selector = event.keys[0];
-        if selector
-            == starknet::core::utils::get_selector_from_name("OwnershipTransferred")
-                .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "OwnershipTransferred")
-                })
-        {
-            let mut key_offset = 0 + 1;
-            let mut data_offset = 0;
-            let previous_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.keys,
-                key_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "previous_owner",
-                            "OwnershipTransferred",
-                            e,
-                        ),
-                    );
-                }
-            };
-            key_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &previous_owner,
-                );
-            let new_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.keys,
-                key_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "new_owner",
-                            "OwnershipTransferred",
-                            e,
-                        ),
-                    );
-                }
-            };
-            key_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &new_owner,
-                );
-            return Ok(
-                Event::OwnableEvent(
-                    Event::OwnershipTransferred(OwnershipTransferred {
-                        previous_owner,
-                        new_owner,
-                    }),
-                ),
-            );
-        }
-        let selector = event.keys[0];
-        if selector
-            == starknet::core::utils::get_selector_from_name("OwnershipTransferStarted")
-                .unwrap_or_else(|_| {
-                    panic!("Invalid selector for {}", "OwnershipTransferStarted")
-                })
-        {
-            let mut key_offset = 0 + 1;
-            let mut data_offset = 0;
-            let previous_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.keys,
-                key_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "previous_owner",
-                            "OwnershipTransferStarted",
-                            e,
-                        ),
-                    );
-                }
-            };
-            key_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &previous_owner,
-                );
-            let new_owner = match cainome::cairo_serde::ContractAddress::cairo_deserialize(
-                &event.keys,
-                key_offset,
-            ) {
-                Ok(v) => v,
-                Err(e) => {
-                    return Err(
-                        format!(
-                            "Could not deserialize field {} for {}: {:?}",
-                            "new_owner",
-                            "OwnershipTransferStarted",
-                            e,
-                        ),
-                    );
-                }
-            };
-            key_offset
-                += cainome::cairo_serde::ContractAddress::cairo_serialized_size(
-                    &new_owner,
-                );
-            return Ok(
-                Event::OwnableEvent(
-                    Event::OwnershipTransferStarted(OwnershipTransferStarted {
-                        previous_owner,
-                        new_owner,
-                    }),
-                ),
-            );
-        }
-        Err(format!("Could not match any event from keys {:?}", event.keys))
-    }
-}
-impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
+impl<A: starknet::accounts::ConnectedAccount + Sync> store<A> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn is_on_whitelist(
+    pub fn pay_oracle_fees(
         &self,
-        element_to_check: &cainome::cairo_serde::ContractAddress,
-    ) -> cainome::cairo_serde::call::FCall<A::Provider, bool> {
+        erc20_address: &cainome::cairo_serde::ContractAddress,
+        amount: &cainome::cairo_serde::U256,
+    ) -> cainome::cairo_serde::call::FCall<A::Provider, ()> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
             .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_check),
+                cainome::cairo_serde::ContractAddress::cairo_serialize(erc20_address),
             );
+        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(amount));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("is_on_whitelist"),
+            entry_point_selector: starknet::macros::selector!("pay_oracle_fees"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn get_whitelist(
+    pub fn compute_final_fee(
         &self,
-    ) -> cainome::cairo_serde::call::FCall<
-        A::Provider,
-        Vec<cainome::cairo_serde::ContractAddress>,
-    > {
+        currency: &cainome::cairo_serde::ContractAddress,
+    ) -> cainome::cairo_serde::call::FCall<A::Provider, cainome::cairo_serde::U256> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
+        __calldata
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("get_whitelist"),
+            entry_point_selector: starknet::macros::selector!("compute_final_fee"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
@@ -851,68 +682,36 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn add_to_whitelist_getcall(
+    pub fn set_final_fee_getcall(
         &self,
-        new_element: &cainome::cairo_serde::ContractAddress,
+        currency: &cainome::cairo_serde::ContractAddress,
+        new_final_fee: &cainome::cairo_serde::U256,
     ) -> starknet::accounts::Call {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(new_element));
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
+        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(new_final_fee));
         starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("add_to_whitelist"),
+            selector: starknet::macros::selector!("set_final_fee"),
             calldata: __calldata,
         }
     }
     #[allow(clippy::ptr_arg)]
-    pub fn add_to_whitelist(
+    pub fn set_final_fee(
         &self,
-        new_element: &cainome::cairo_serde::ContractAddress,
+        currency: &cainome::cairo_serde::ContractAddress,
+        new_final_fee: &cainome::cairo_serde::U256,
     ) -> starknet::accounts::Execution<A> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
-            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(new_element));
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
+        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(new_final_fee));
         let __call = starknet::accounts::Call {
             to: self.address,
-            selector: starknet::macros::selector!("add_to_whitelist"),
-            calldata: __calldata,
-        };
-        self.account.execute(vec![__call])
-    }
-    #[allow(clippy::ptr_arg)]
-    #[allow(clippy::too_many_arguments)]
-    pub fn remove_from_whitelist_getcall(
-        &self,
-        element_to_remove: &cainome::cairo_serde::ContractAddress,
-    ) -> starknet::accounts::Call {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_remove),
-            );
-        starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("remove_from_whitelist"),
-            calldata: __calldata,
-        }
-    }
-    #[allow(clippy::ptr_arg)]
-    pub fn remove_from_whitelist(
-        &self,
-        element_to_remove: &cainome::cairo_serde::ContractAddress,
-    ) -> starknet::accounts::Execution<A> {
-        use cainome::cairo_serde::CairoSerde;
-        let mut __calldata = vec![];
-        __calldata
-            .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_remove),
-            );
-        let __call = starknet::accounts::Call {
-            to: self.address,
-            selector: starknet::macros::selector!("remove_from_whitelist"),
+            selector: starknet::macros::selector!("set_final_fee"),
             calldata: __calldata,
         };
         self.account.execute(vec![__call])
@@ -972,39 +771,41 @@ impl<A: starknet::accounts::ConnectedAccount + Sync> address_whitelist<A> {
         self.account.execute(vec![__call])
     }
 }
-impl<P: starknet::providers::Provider + Sync> address_whitelistReader<P> {
+impl<P: starknet::providers::Provider + Sync> storeReader<P> {
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn is_on_whitelist(
+    pub fn pay_oracle_fees(
         &self,
-        element_to_check: &cainome::cairo_serde::ContractAddress,
-    ) -> cainome::cairo_serde::call::FCall<P, bool> {
+        erc20_address: &cainome::cairo_serde::ContractAddress,
+        amount: &cainome::cairo_serde::U256,
+    ) -> cainome::cairo_serde::call::FCall<P, ()> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
         __calldata
             .extend(
-                cainome::cairo_serde::ContractAddress::cairo_serialize(element_to_check),
+                cainome::cairo_serde::ContractAddress::cairo_serialize(erc20_address),
             );
+        __calldata.extend(cainome::cairo_serde::U256::cairo_serialize(amount));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("is_on_whitelist"),
+            entry_point_selector: starknet::macros::selector!("pay_oracle_fees"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
     }
     #[allow(clippy::ptr_arg)]
     #[allow(clippy::too_many_arguments)]
-    pub fn get_whitelist(
+    pub fn compute_final_fee(
         &self,
-    ) -> cainome::cairo_serde::call::FCall<
-        P,
-        Vec<cainome::cairo_serde::ContractAddress>,
-    > {
+        currency: &cainome::cairo_serde::ContractAddress,
+    ) -> cainome::cairo_serde::call::FCall<P, cainome::cairo_serde::U256> {
         use cainome::cairo_serde::CairoSerde;
         let mut __calldata = vec![];
+        __calldata
+            .extend(cainome::cairo_serde::ContractAddress::cairo_serialize(currency));
         let __call = starknet::core::types::FunctionCall {
             contract_address: self.address,
-            entry_point_selector: starknet::macros::selector!("get_whitelist"),
+            entry_point_selector: starknet::macros::selector!("compute_final_fee"),
             calldata: __calldata,
         };
         cainome::cairo_serde::call::FCall::new(__call, self.provider())
