+++ scripts/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod bind;
+pub mod deploy;
+pub mod types;
+pub mod utils;
